+++ lib/posthog/feature_flags.rb
@@ -10,8 +10,12 @@ class PostHog
   class InconclusiveMatchError < StandardError
   end
 
+  class DecideAPIError < StandardError
+  end
+
   class FeatureFlagsPoller
     include PostHog::Logging
+    include PostHog::Utils
 
     def initialize(polling_interval, personal_api_key, project_api_key, host)
       @polling_interval = polling_interval || 30
@@ -55,11 +59,10 @@ class PostHog
 
       decide_data = _request_feature_flag_evaluation(request_data)
 
-      if !decide_data.key?('featureFlags')
-        raise StandardError.new(decide_data.to_json)
+      if !decide_data.key?(:featureFlags)
+        raise DecideAPIError.new(decide_data.to_json)
       else
-        feature_variants = decide_data["featureFlags"] || {}
-        return feature_variants
+        decide_data[:featureFlags] || {}
       end
     end
 
@@ -67,11 +70,19 @@ class PostHog
       # make sure they're loaded on first run
       load_feature_flags
 
+      symbolize_keys! groups
+      symbolize_keys! person_properties
+      symbolize_keys! group_properties
+
+      group_properties.each do |key, value|
+        symbolize_keys! value
+      end
+
       response = nil
       feature_flag = nil
 
       @feature_flags.each do |flag|
-        if key == flag['key']
+        if key == flag[:key]
           feature_flag = flag
           break
         end
@@ -84,7 +95,7 @@ class PostHog
         rescue InconclusiveMatchError => e
           logger.debug "Failed to compute flag #{key} locally: #{e}"
         rescue StandardError => e
-          logger.error "Error computing flag locally: #{e}"
+          logger.error "Error computing flag locally: #{e}. #{e.backtrace.join("\n")}"
         end
       end
 
@@ -94,7 +105,7 @@ class PostHog
           response = flags[key]
           logger.debug "Successfully computed flag remotely: #{key} -> #{response}"
         rescue StandardError => e
-          logger.error "Error computing flag remotely: #{e}"
+          logger.error "Error computing flag remotely: #{e}. #{e.backtrace.join("\n")}"
           response = default_result
         end
       end
@@ -111,7 +122,7 @@ class PostHog
 
       @feature_flags.each do |flag|
         begin
-          response[flag['key']] = _compute_flag_locally(flag, distinct_id, groups, person_properties, group_properties)
+          response[flag[:key]] = _compute_flag_locally(flag, distinct_id, groups, person_properties, group_properties)
         rescue InconclusiveMatchError => e
           fallback_to_decide = true
         rescue StandardError => e
@@ -128,7 +139,7 @@ class PostHog
           logger.error "Error computing flag remotely: #{e}"
         end
       end
-      return response
+      response
     end
 
     def shutdown_poller()
@@ -140,12 +151,15 @@ class PostHog
     def self.match_property(property, property_values)
       # only looks for matches where key exists in property_values
       # doesn't support operator is_not_set
+      
+      PostHog::Utils.symbolize_keys! property
+      PostHog::Utils.symbolize_keys! property_values
 
-      key = property['key']
-      value = property['value']
-      operator = property['operator'] || 'exact'
+      key = property[:key].to_sym
+      value = property[:value]
+      operator = property[:operator] || 'exact'
 
-      if !property_values.has_key?(key)
+      if !property_values.key?(key)
         raise InconclusiveMatchError.new("Property #{key} not found in property_values")
       elsif operator == 'is_not_set'
         raise InconclusiveMatchError.new("Operator is_not_set not supported")
@@ -153,80 +167,68 @@ class PostHog
 
       override_value = property_values[key]
 
-      if operator == 'exact'
-        if value.is_a?(Array)
-          return value.include?(override_value)
-        end
-        return value == override_value
-      elsif operator == 'is_not'
-        if value.is_a?(Array)
-          return !value.include?(override_value)
-        end
-        return value != override_value
-      elsif operator == 'is_set'
-        return property_values.has_key?(key)
-      elsif operator == 'icontains'
-        return override_value.to_s.downcase.include?(value.to_s.downcase)
-      elsif operator == 'not_icontains'
-        return !override_value.to_s.downcase.include?(value.to_s.downcase)
-      elsif operator == 'regex'
-        return is_valid_regex(value.to_s) && !Regexp.new(value.to_s).match(override_value.to_s).nil?
-      elsif operator == 'not_regex'
-        return is_valid_regex(value.to_s) && Regexp.new(value.to_s).match(override_value.to_s).nil?
-      elsif operator == 'gt'
-        return override_value.class == value.class && override_value > value
-      elsif operator == 'gte'
-        return override_value.class == value.class && override_value >= value
-      elsif operator == 'lt'
-        return override_value.class == value.class && override_value < value
-      elsif operator == 'lte'
-        return override_value.class == value.class && override_value <= value
+      case operator
+      when 'exact'
+        value.is_a?(Array) ? value.include?(override_value) : value == override_value
+      when 'is_not'
+        value.is_a?(Array) ? !value.include?(override_value) : value != override_value
+      when'is_set'
+        property_values.key?(key)
+      when 'icontains'
+        override_value.to_s.downcase.include?(value.to_s.downcase)
+      when 'not_icontains'
+        !override_value.to_s.downcase.include?(value.to_s.downcase)
+      when 'regex'
+        PostHog::Utils.is_valid_regex(value.to_s) && !Regexp.new(value.to_s).match(override_value.to_s).nil?
+      when 'not_regex'
+        PostHog::Utils.is_valid_regex(value.to_s) && Regexp.new(value.to_s).match(override_value.to_s).nil?
+      when 'gt'
+        override_value.class == value.class && override_value > value
+      when 'gte'
+        override_value.class == value.class && override_value >= value
+      when 'lt'
+        override_value.class == value.class && override_value < value
+      when 'lte'
+        override_value.class == value.class && override_value <= value
       else
-        return false
-      end
-
-    end
-
-    def self.is_valid_regex(regex)
-      begin
-        Regexp.new(regex)
-        return true
-      rescue RegexpError
-        return false
+        logger.error "Unknown operator: #{operator}"
+        false
       end
     end
 
     private
 
     def _compute_flag_locally(flag, distinct_id, groups = {}, person_properties = {}, group_properties = {})
-      if flag['ensure_experience_continuity']
+      if flag[:ensure_experience_continuity]
         raise InconclusiveMatchError.new("Flag has experience continuity enabled")
       end
 
-      if !flag['active']
-        return false
-      end
+      return false if !flag[:active]
 
-      flag_filters = flag['filters'] || {}
-      aggregation_group_type_index = flag_filters['aggregation_group_type_index']
+      flag_filters = flag[:filters] || {}
+      symbolize_keys! flag_filters
+
+      aggregation_group_type_index = flag_filters[:aggregation_group_type_index]
       if !aggregation_group_type_index.nil?
-        group_name = @group_type_mapping[aggregation_group_type_index.to_s]
+        group_name = @group_type_mapping[aggregation_group_type_index.to_s.to_sym]
 
         if group_name.nil?
-          logger.warn "[FEATURE FLAGS] Unknown group type index #{aggregation_group_type_index} for feature flag #{flag['key']}"
+          logger.warn "[FEATURE FLAGS] Unknown group type index #{aggregation_group_type_index} for feature flag #{flag[:key]}"
           # failover to `/decide/`
           raise InconclusiveMatchError.new("Flag has unknown group type index")
         end
 
-        if !groups.has_key?(group_name)
+        group_name_symbol = group_name.to_sym
+
+        if !groups.key?(group_name_symbol)
           # Group flags are never enabled if appropriate `groups` aren't passed in
           # don't failover to `/decide/`, since response will be the same
-          logger.warn "[FEATURE FLAGS] Can't compute group feature flag: #{flag['key']} without group names passed in"
+          logger.warn "[FEATURE FLAGS] Can't compute group feature flag: #{flag[:key]} without group names passed in"
           return false
         end
 
-        focused_group_properties = group_properties[group_name]
-        return match_feature_flag_properties(flag, groups[group_name], focused_group_properties)
+        focused_group_properties = group_properties[group_name_symbol]
+        return match_feature_flag_properties(flag, groups[group_name_symbol], focused_group_properties)
       else
         return match_feature_flag_properties(flag, distinct_id, person_properties)
       end
@@ -234,8 +236,10 @@ class PostHog
     end
 
     def match_feature_flag_properties(flag, distinct_id, properties)
-      flag_filters = flag['filters'] || {}
-      flag_conditions = flag_filters['groups'] || []
+      flag_filters = flag[:filters] || {}
+      symbolize_keys! flag_filters
+
+      flag_conditions = flag_filters[:groups] || []
       is_inconclusive = false
       result = nil
 
@@ -256,16 +260,17 @@ class PostHog
         raise InconclusiveMatchError.new("Can't determine if feature flag is enabled or not with given properties")
       end
 
-      # We can only return False when either all conditions are False, or
-      # no condition was inconclusive.
+      # We can only return False when all conditions are False
       return false
     end
 
     def is_condition_match(flag, distinct_id, condition, properties)
-      rollout_percentage = condition['rollout_percentage']
+      symbolize_keys! condition
+      rollout_percentage = condition[:rollout_percentage]
 
-      if !(condition['properties'] || []).empty?
-        if !condition['properties'].all? { |prop|
+      if !(condition[:properties] || []).empty?
+        if !condition[:properties].all? { |prop|
+            symbolize_keys! prop
             FeatureFlagsPoller.match_property(prop, properties)
           }
           return false
@@ -274,7 +279,7 @@ class PostHog
         end
       end
 
-      if !rollout_percentage.nil? and _hash(flag['key'], distinct_id) > (rollout_percentage.to_f/100)
+      if !rollout_percentage.nil? and _hash(flag[:key], distinct_id) > (rollout_percentage.to_f/100)
         return false
       end
       
@@ -291,24 +296,27 @@ class PostHog
     end
 
     def get_matching_variant(flag, distinct_id)
-      hash_value = _hash(flag['key'], distinct_id, salt="variant")
-      variant_lookup_table(flag).each do |variant|
-        if (
-           hash_value >= variant[:value_min] and hash_value <= variant[:value_max]
-        )
-          return variant[:key]
-        end
-      end
-      return nil
+      hash_value = _hash(flag[:key], distinct_id, salt="variant")
+      matching_variant = variant_lookup_table(flag).find { |variant|
+          hash_value >= variant[:value_min] and hash_value < variant[:value_max]
+      }
+      matching_variant.nil? ? nil : matching_variant[:key]
     end
 
     def variant_lookup_table(flag)
       lookup_table = []
       value_min = 0
-      multivariates = ((flag['filters'] || {})['multivariate'] || {})['variants'] || []
+      flag_filters = flag[:filters] || {}
+      symbolize_keys! flag_filters
+
+      variants = flag_filters[:multivariate] || {}
+      symbolize_keys! variants
+
+      multivariates = variants[:variants] || []
       multivariates.each do |variant|
-        value_max = value_min + variant['rollout_percentage'].to_f / 100
-        lookup_table << {'value_min': value_min, 'value_max': value_max, 'key': variant['key']}
+        symbolize_keys! variant
+        value_max = value_min + variant[:rollout_percentage].to_f / 100
+        lookup_table << {'value_min': value_min, 'value_max': value_max, 'key': variant[:key]}
         value_min = value_max
       end
       return lookup_table
@@ -318,11 +326,17 @@ class PostHog
       res = _request_feature_flag_definitions
       @feature_flags.clear
 
-      if !res.key?('flags')
+      if !res.key?(:flags)
         logger.error "Failed to load feature flags: #{res}"
       else
-        @feature_flags = res['flags'] || []
-        @group_type_mapping = res['group_type_mapping'] || {}
+        @feature_flags = res[:flags] || []
+        @feature_flags.each do |flag|
+          symbolize_keys! flag
+        end
+        
+        @group_type_mapping = res[:group_type_mapping] || {}
+        symbolize_keys! @group_type_mapping
+
         logger.debug "Loaded #{@feature_flags.length} feature flags"
         if @loaded_flags_successfully_once.false?
           @loaded_flags_successfully_once.make_true
@@ -336,7 +350,7 @@ class PostHog
       req['Authorization'] = "Bearer #{@personal_api_key}"
       req['User-Agent'] = "posthog-ruby#{PostHog::VERSION}"
 
-      return _request(uri, req)
+      _request(uri, req)
     end
 
     def _request_feature_flag_evaluation(data={})
@@ -346,7 +360,7 @@ class PostHog
       data['token'] = @project_api_key
       req.body = data.to_json
 
-      return _request(uri, req)
+      _request(uri, req)
     end
 
     def _request(uri, request_object)
@@ -357,7 +371,7 @@ class PostHog
         res_body = nil
         Net::HTTP.start(uri.hostname, uri.port, use_ssl: uri.scheme == 'https') do |http|
           res = http.request(request_object)
-          return JSON.parse(res.body)
+          symbolize_keys! JSON.parse(res.body)
         end
       rescue Timeout::Error,
              Errno::EINVAL,